+++ authentication-tokens/src/main/java/net/optionfactory/spring/authentication/tokens/jwt/JweAuthenticationConfigurer.java
@@ -18,7 +18,6 @@ import java.security.interfaces.ECPrivateKey;
 import java.security.interfaces.ECPublicKey;
 import java.security.interfaces.RSAPublicKey;
 import java.util.List;
-import java.util.Locale;
 import javax.crypto.SecretKey;
 import net.optionfactory.spring.authentication.tokens.HeaderAndScheme;
 import net.optionfactory.spring.authentication.tokens.jwt.JwtTokenProcessor.JweProcessor;
@@ -131,7 +130,7 @@ public interface JweAuthenticationConfigurer extends JwtAuthenticationConfigurer
         public Builder matchHeader(String header, String authScheme) {
             Assert.notNull(header, "header cannot be null");
             Assert.notNull(authScheme, "authScheme cannot be null");
-            this.hs = new HeaderAndScheme(header, authScheme.toUpperCase(Locale.ROOT).trim() + " ");
+            this.hs = new HeaderAndScheme(header, authScheme);
             return this;
         }